requirements.txt
fastapi
uvicorn
sqlalchemy
sqladmin
wtforms
itsdangerous
openai
python-dotenv
apscheduler
python-multipart
aiogram
